CONT, FLOR, NOTE fixes; DOOR, MISC; NetImmerse types
Fixed the type checking for FLOR/PFIG.

Fixed clone and clear code for NOTE.

Implemented the DOOR and MISC form types, and finished and fixed the CONT form type.

Implemented the BSLeafAnimNode, BSLODTriShape, and NiSwitchNode NetImmerse block types. Fixed flags width on NiTimeController.

+++ native/src/DovahKit/nif/blocks/NiTimeController.cpp
@@ -7,7 +7,7 @@ namespace nifDK::block_types {
    void NiTimeController::parse(file_reader& reader) {
       reader.read_ref(this->next);
       {
-         uint32_t flags = 0;
+         uint16_t flags = 0;
          reader.read(flags);
          this->config.anim_type  = (animation_type)(flags & 1);
          this->config.cycle_type = (cycle_type)((flags >> 1) & 0b11);
